changed 'unsigned long int' definitions to 'unsigned int' (continued)

+++ drivers/block/ata.c
@@ -901,7 +901,7 @@ int ata_write(__dev_t dev, __blk_t block, char *buffer, int blksize)
 	return drive->fsop->write_block(dev, block, buffer, blksize);
 }
 
-int ata_ioctl(struct inode *i, int cmd, unsigned long int arg)
+int ata_ioctl(struct inode *i, int cmd, unsigned int arg)
 {
 	struct ide *ide;
 	struct ata_drv *drive;
