+++ drivers/block/floppy.c
@@ -749,7 +749,7 @@ int fdc_write(__dev_t dev, __blk_t block, char *buffer, int blksize)
 	return sectors_written * BPS;
 }
 
-int fdc_ioctl(struct inode *i, int cmd, unsigned long int arg)
+int fdc_ioctl(struct inode *i, int cmd, unsigned int arg)
 {
 	unsigned char minor;
 	struct hd_geometry *geom;
