+++ include/fiwix/atapi_cd.h
@@ -15,7 +15,7 @@
 int atapi_cd_open(struct inode *, struct fd *);
 int atapi_cd_close(struct inode *, struct fd *);
 int atapi_cd_read(__dev_t, __blk_t, char *, int);
-int atapi_cd_ioctl(struct inode *, int, unsigned long int);
+int atapi_cd_ioctl(struct inode *, int, unsigned int);
 __loff_t atapi_cd_llseek(struct inode *, __loff_t);
 int atapi_cd_init(struct ide *, struct ata_drv *);
